Block adding dependencies/hierarchies that duplicate dependencies
Block adding any dependencies/hierarchies that would lead to there being
multiple dependencies between one hierarchy and another. For example, if
A is a supertask of B, and C is a supertask of D, and A is a dependee of
C, then adding a dependency from B to D is redundant.