Fix incorrect task labels bug
The index returned by the mouse-move event only specifies the index of
the artist it's within. Therefore when tasks were deleted, there was a
mismatch between the index number and the task ID. This was addressed
by creating a list of nodes which mirrored the PathCollection, then
looking up the node-ID based on the index.