Extract new NetworkGraph class
Currently TaskSystem was doing too much work, and attributes didn't make
sense when getting upstream/downstream subgraphs. Added an extra layer
to handle these responsibilities. Also expanded the number of methods on
the Views and IViews; lots of chaining to get to the object you want,
but cuts down on repetition as calls are passed down the chain. Finally,
I combined the dfs and wfs search variants because I thought it looked
nicer.