Add logging
Tried to keep things simple; log user actions, any failures, and calls
across layers. DEBUG for read-only methods and INFO for those that
change state. WARNING for failed logic-layer methods (as exceptions are
used for communication to the GUI) and ERROR for failed data-layer
methods (much harder to recover from that).