Add caching data layer
Was annoying me that I was loading the next_unused_task_uid file all the
time, so decided to cache it. Also realised that I'd been gitignoring
all of my data layer implementations. So there's a bunch of data-layer
code in this commit that I thought I'd commited in earlier. Oops.