Introduce graph allowing loops and subgraph views
A lot in this commit. Introducing a graph that allowed self loops was
done to remove a few responsibilities from the Simple DiGraph class,
which I felt was getting a bit overful.

The subgraph data structures should theoretically improve performance
for actions such as:
- Checking if a node is in a subgraph
- Checking if an edge is in a subgraph
Ironically, running the test suite after the change actually takes
longer - I'm going to chose to ignore that.
This change is only half-done, as I still need to allow for
multi-starting-node subgraphs.