Failed attempt to update tree-views on task selection
Thought it'd be nice for all tree-views to be focused on the currently
selected task (like how the graphs show the current task with a red
node). This proved to be a tkinter rabbit hole, and I also uncovered a
bug that I'd unintentionally been suppressing. What a waste of time.