Layout task network diagram
Attempts to visualise the task network in the 'Task Network' tab.
Hierarchies & dependencies are both shown, with tasks represented as
cylinders. Hovering over a cylinder will reveal the task name, and
clicking on it will cause the selected task to appear in the context
pane. Heuristic approach used to find a network layout that minimises
edge crossover and distance.

Really quite proud of this.