Create unconstrained network graph superclass
There's a good reason for this, honest. The network graph drawing code
works whether dependency duplication and crossovers are allowed or not.
When a user introduces a duplication or crossover, it's hard to
visualise the problem. Therefore by allowing the graph to be drawn
with the duplication/crossover in place allows the problem to be
communicated far more easily. By splitting out the 'unconstrained'
network graph into a superclass, I can allow this to be passed to the
graph drawing code, while still maintaining the 'constrained' subclass
for general use.